update fixed savings services
