+++ src/services/fixedsavings/index.ts
@@ -1,0 +1,6 @@
+import { AddSavings } from './AddSavings'
+import { SavingEstimate } from './SavingEstimate'
+import { FsSaves } from './FsSaves'
+import { FsInfo } from './FsInfo'
+import { WithdrawSavings } from './WithdrawSavings'
+export { AddSavings, SavingEstimate, FsSaves, FsInfo, WithdrawSavings }
